requirements.txt
astroid==1.6.3
autopep8==1.3.5
cycler==0.10.0
demjson==2.2.4
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==2.2.2
mccabe==0.6.1
mysql-connector-python==8.0.11
numpy==1.14.3
protobuf==3.5.2.post1
psutil==5.4.5
pycodestyle==2.4.0
pylint==1.8.4
pymongo==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
six==1.11.0
wrapt==1.10.11
